+++ app/google-services.json
@@ -10,7 +10,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:246301202402:android:db7002f4295e49e4b84bbb",
         "android_client_info": {
-          "package_name": "edu.psu.cto5068.like_birbs"
+          "package_name": "ga.hnbenterprises.like_birbs"
         }
       },
       "oauth_client": [
