+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="edu.psu.cto5068.like_birbs">
+    package="ga.hnbenterprises.like_birbs">
 
     <uses-permission android:name="android.permission.VIBRATE" />
