--- app/src/main/java/edu/psu/cto5068/like_birbs/Game_over.java
@@ -1,265 +1,0 @@
-package edu.psu.cto5068.like_birbs;
-
-import android.annotation.SuppressLint;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.preference.PreferenceManager;
-
-import android.content.Context;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.os.Handler;
-import android.os.Vibrator;
-import android.util.Log;
-import android.view.MotionEvent;
-import android.view.View;
-
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * An example full-screen activity that shows and hides the system UI (i.e.
- * status bar and navigation/system bar) with user interaction.
- */
-public class Game_over extends AppCompatActivity {
-
-    private static final String USER_KEY = "username";
-    private static final String ENV_KEY = "env";
-    private static final String SCORE_KEY = "score";
-    private static final String TAG = "Game_over";
-    private FirebaseFirestore db = FirebaseFirestore.getInstance();
-    /**
-     * Whether or not the system UI should be auto-hidden after
-     * {@link #AUTO_HIDE_DELAY_MILLIS} milliseconds.
-     */
-    private static final boolean AUTO_HIDE = true;
-
-    /**
-     * If {@link #AUTO_HIDE} is set, the number of milliseconds to wait after
-     * user interaction before hiding the system UI.
-     */
-    private static final int AUTO_HIDE_DELAY_MILLIS = 3000;
-
-    /**
-     * Some older devices needs a small delay between UI widget updates
-     * and a change of the status and navigation bar.
-     */
-    private static final int UI_ANIMATION_DELAY = 300;
-    private final Handler mHideHandler = new Handler();
-    private View mContentView;
-    private boolean keepMusicGoing = false;
-    private final Runnable mHidePart2Runnable = new Runnable() {
-        @SuppressLint("InlinedApi")
-        @Override
-        public void run() {
-            // Delayed removal of status and navigation bar
-
-            // Note that some of these constants are new as of API 16 (Jelly Bean)
-            // and API 19 (KitKat). It is safe to use them, as they are inlined
-            // at compile-time and do nothing on earlier devices.
-            mContentView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LOW_PROFILE
-                    | View.SYSTEM_UI_FLAG_FULLSCREEN
-                    | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                    | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
-        }
-    };
-    private View mControlsView;
-    private final Runnable mShowPart2Runnable = new Runnable() {
-        @Override
-        public void run() {
-            // Delayed display of UI elements
-            ActionBar actionBar = getSupportActionBar();
-            if (actionBar != null) {
-                actionBar.show();
-            }
-            mControlsView.setVisibility(View.VISIBLE);
-        }
-    };
-    private boolean mVisible;
-    private final Runnable mHideRunnable = new Runnable() {
-        @Override
-        public void run() {
-            hide();
-        }
-    };
-    /**
-     * Touch listener to use for in-layout UI controls to delay hiding the
-     * system UI. This is to prevent the jarring behavior of controls going away
-     * while interacting with activity UI.
-     */
-    private final View.OnTouchListener mDelayHideTouchListener = new View.OnTouchListener() {
-        @Override
-        public boolean onTouch(View view, MotionEvent motionEvent) {
-            if (AUTO_HIDE) {
-                delayedHide(AUTO_HIDE_DELAY_MILLIS);
-            }
-            return false;
-        }
-    };
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_game_over);
-        Intent intent = getIntent();
-        Bundle extras = intent.getExtras();
-        Integer env = extras.getInt("env");
-        Integer score = extras.getInt("totalGens");
-
-        String environment;
-        switch (env) {
-            case 1 :
-                environment = "Meadow";
-                break;
-            case 2:
-                environment = "Desert";
-                break;
-            case 3:
-                environment = "Snow";
-                break;
-            case 4:
-                environment = "Island";
-                break;
-            default:
-                environment = "Unknown";
-        }
-
-        SharedPreferences sharedPreferences =
-                PreferenceManager.getDefaultSharedPreferences(this /* Activity context */);
-        String username = sharedPreferences.getString("username", "");
-        boolean upload_score = sharedPreferences.getBoolean("upload_score", true);
-
-        if (upload_score) {
-            Map<String, Object> dataToSave = new HashMap<String, Object>();
-            dataToSave.put(USER_KEY, username);
-            dataToSave.put(ENV_KEY, environment);
-            dataToSave.put(SCORE_KEY, score);
-
-            SharedPreferences sharedPref = this.getSharedPreferences("highscoreCache",Context.MODE_PRIVATE);
-            final SharedPreferences.Editor editor = sharedPref.edit();
-
-            db.collection("highscores")
-                    .add(dataToSave)
-                    .addOnSuccessListener(new OnSuccessListener<DocumentReference>() {
-                        @Override
-                        public void onSuccess(DocumentReference documentReference) {
-                            editor.putString("highscoreRef", documentReference.getId());
-                            editor.apply();
-                            Log.d(TAG, "DocumentSnapshot written with ID: " + documentReference.getId());
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Log.w(TAG, "Error adding document", e);
-                        }
-                    });
-        }
-
-        mVisible = true;
-    }
-
-    @Override
-    protected void onPostCreate(Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
-
-        // Trigger the initial hide() shortly after the activity has been
-        // created, to briefly hint to the user that UI controls
-        // are available.
-        delayedHide(100);
-    }
-
-    private void toggle() {
-        if (mVisible) {
-            hide();
-        } else {
-            show();
-        }
-    }
-
-    private void hide() {
-        // Hide UI first
-        ActionBar actionBar = getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.hide();
-        }
-        mVisible = false;
-
-    }
-
-    @SuppressLint("InlinedApi")
-    private void show() {
-        // Show the system bar
-        mContentView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION);
-        mVisible = true;
-
-        // Schedule a runnable to display UI elements after a delay
-        mHideHandler.removeCallbacks(mHidePart2Runnable);
-        mHideHandler.postDelayed(mShowPart2Runnable, UI_ANIMATION_DELAY);
-    }
-
-    /**
-     * Schedules a call to hide() in delay milliseconds, canceling any
-     * previously scheduled calls.
-     */
-    private void delayedHide(int delayMillis) {
-        mHideHandler.removeCallbacks(mHideRunnable);
-        mHideHandler.postDelayed(mHideRunnable, delayMillis);
-    }
-
-    public void gameClick(View view) {
-        int id = view.getId();
-        Vibrator v = (Vibrator) getSystemService(Context.VIBRATOR_SERVICE);
-
-        switch (id) {
-            case (R.id.goToScoresButton):
-                // code to view scores
-                v.vibrate(100);
-                Intent nextScreenIntent = new Intent(this, Leaderbirb.class);
-                keepMusicGoing = true;
-                startActivity(nextScreenIntent);
-                break;
-            case (R.id.goToMenuButton):
-                // launch dialog asking to save / no save and quit
-                v.vibrate(100);
-                Intent menuIntent = new Intent(this, FullscreenActivity.class);
-                menuIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                keepMusicGoing = true;
-                startActivity(menuIntent);
-                break;
-            default:
-                // error
-                break;
-        }
-    }
-
-    public void onResume() {
-        super.onResume();
-        keepMusicGoing = false;
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        boolean playMusicPreference = sharedPreferences.getBoolean("play_music", true);
-
-        if(playMusicPreference) {
-            startService(new Intent(Game_over.this, SoundService.class));
-        }
-    }
-    @Override
-    public void onPause(){
-        super.onPause();
-        if (!keepMusicGoing) {
-            stopService(new Intent(Game_over.this, SoundService.class));
-        }
-    }
-}
