--- app/src/main/java/edu/psu/cto5068/like_birbs/Leaderbirb.java
@@ -1,229 +1,0 @@
-package edu.psu.cto5068.like_birbs;
-
-import android.annotation.SuppressLint;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.preference.PreferenceManager;
-
-import android.content.Context;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.os.Handler;
-import android.util.Log;
-import android.view.MotionEvent;
-import android.view.View;
-import android.widget.ListView;
-
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.EventListener;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.google.firebase.firestore.Query;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import edu.psu.cto5068.like_birbs.leaderbirb.HighScore;
-import edu.psu.cto5068.like_birbs.leaderbirb.ScoreListAdapter;
-
-/**
- * An example full-screen activity that shows and hides the system UI (i.e.
- * status bar and navigation/system bar) with user interaction.
- */
-public class Leaderbirb extends AppCompatActivity {
-
-    private static final String USER_KEY = "username";
-    private static final String SCORE_KEY = "score";
-    private static final String TAG = "Leaderbirb";
-    private boolean keepMusicGoing = false;
-    private FirebaseFirestore db = FirebaseFirestore.getInstance();
-
-    /**
-     * Whether or not the system UI should be auto-hidden after
-     * {@link #AUTO_HIDE_DELAY_MILLIS} milliseconds.
-     */
-    private static final boolean AUTO_HIDE = true;
-
-    /**
-     * If {@link #AUTO_HIDE} is set, the number of milliseconds to wait after
-     * user interaction before hiding the system UI.
-     */
-    private static final int AUTO_HIDE_DELAY_MILLIS = 3000;
-
-    /**
-     * Some older devices needs a small delay between UI widget updates
-     * and a change of the status and navigation bar.
-     */
-    private static final int UI_ANIMATION_DELAY = 300;
-    private final Handler mHideHandler = new Handler();
-    private View mContentView;
-    private final Runnable mHidePart2Runnable = new Runnable() {
-        @SuppressLint("InlinedApi")
-        @Override
-        public void run() {
-            // Delayed removal of status and navigation bar
-
-            // Note that some of these constants are new as of API 16 (Jelly Bean)
-            // and API 19 (KitKat). It is safe to use them, as they are inlined
-            // at compile-time and do nothing on earlier devices.
-            mContentView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LOW_PROFILE
-                    | View.SYSTEM_UI_FLAG_FULLSCREEN
-                    | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                    | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
-        }
-    };
-    private View mControlsView;
-    private final Runnable mShowPart2Runnable = new Runnable() {
-        @Override
-        public void run() {
-            // Delayed display of UI elements
-            ActionBar actionBar = getSupportActionBar();
-            if (actionBar != null) {
-                actionBar.show();
-            }
-            mControlsView.setVisibility(View.VISIBLE);
-        }
-    };
-    private boolean mVisible;
-    private final Runnable mHideRunnable = new Runnable() {
-        @Override
-        public void run() {
-            hide();
-        }
-    };
-    /**
-     * Touch listener to use for in-layout UI controls to delay hiding the
-     * system UI. This is to prevent the jarring behavior of controls going away
-     * while interacting with activity UI.
-     */
-    private final View.OnTouchListener mDelayHideTouchListener = new View.OnTouchListener() {
-        @Override
-        public boolean onTouch(View view, MotionEvent motionEvent) {
-            if (AUTO_HIDE) {
-                delayedHide(AUTO_HIDE_DELAY_MILLIS);
-            }
-            return false;
-        }
-    };
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        setContentView(R.layout.activity_leaderbirb);
-
-        mVisible = true;
-
-        final ListView mListView = findViewById(R.id.listView);
-        final Context mContext = this;
-
-        SharedPreferences sharedPref = this.getSharedPreferences("highscoreCache", Context.MODE_PRIVATE);
-        final String highscoreRef = sharedPref.getString("highscoreRef", "");
-
-        db.collection("highscores")
-                .orderBy("score", Query.Direction.DESCENDING)
-                .addSnapshotListener(new EventListener<QuerySnapshot>() {
-                    @Override
-                    public void onEvent(@Nullable QuerySnapshot value,
-                                        @Nullable FirebaseFirestoreException e) {
-                        if (e != null) {
-                            Log.w(TAG, "Listen failed.", e);
-                            return;
-                        }
-
-                        SharedPreferences sharedPreferences =
-                                PreferenceManager.getDefaultSharedPreferences(mContext /* Activity context */);
-                        int num_scores = Integer.parseInt(sharedPreferences.getString("num_scores", ""));
-
-
-
-                        List<HighScore> highscores = new ArrayList<>();
-                        int rank = 1;
-                        for (QueryDocumentSnapshot doc : value) {
-                            if (doc.get("username") != null && doc.get("env") != null && doc.get("score") != null && highscores.size() < num_scores) {
-                                highscores.add(new HighScore(rank, doc.getString("username"), doc.getString("env"), doc.getLong("score").intValue(), (!highscoreRef.equals("") && highscoreRef.equals(doc.getId()))));
-                                rank++;
-                            }
-                        }
-                        Log.d(TAG, "Current highscores: " + highscores);
-                        ScoreListAdapter adapter = new ScoreListAdapter(mContext, R.layout.leaderbirb_adapter_view, highscores);
-                        mListView.setAdapter(adapter);
-                    }
-                });
-    }
-
-    @Override
-    protected void onPostCreate(Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
-
-        // Trigger the initial hide() shortly after the activity has been
-        // created, to briefly hint to the user that UI controls
-        // are available.
-        delayedHide(100);
-    }
-
-    private void toggle() {
-        if (mVisible) {
-            hide();
-        } else {
-            show();
-        }
-    }
-
-    private void hide() {
-        // Hide UI first
-        ActionBar actionBar = getSupportActionBar();
-        if (actionBar != null) {
-            actionBar.hide();
-        }
-        mVisible = false;
-
-    }
-
-    @SuppressLint("InlinedApi")
-    private void show() {
-        mVisible = true;
-
-    }
-
-    /**
-     * Schedules a call to hide() in delay milliseconds, canceling any
-     * previously scheduled calls.
-     */
-    private void delayedHide(int delayMillis) {
-
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        keepMusicGoing = false;
-        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this);
-        boolean playMusicPreference = sharedPreferences.getBoolean("play_music", true);
-
-        if(playMusicPreference) {
-            startService(new Intent(Leaderbirb.this, SoundService.class));
-        }
-    }
-    @Override
-    public void onPause(){
-        super.onPause();
-        if (!keepMusicGoing) {
-            stopService(new Intent(Leaderbirb.this, SoundService.class));
-        }
-    }
-}
