--- app/src/main/java/edu/psu/cto5068/like_birbs/LogMsg.java
@@ -1,17 +1,0 @@
-package edu.psu.cto5068.like_birbs;
-
-public class LogMsg {
-    private String msg;
-
-    public LogMsg(String msg) {
-        this.msg = msg;
-    }
-
-    public String getMsg() {
-        return msg;
-    }
-
-    public void setMsg(String msg) {
-        this.msg = msg;
-    }
-}
