--- app/src/main/java/edu/psu/cto5068/like_birbs/RandomEvent.java
@@ -1,4 +1,0 @@
-package edu.psu.cto5068.like_birbs;
-
-public class RandomEvent {
-}
