--- app/src/main/java/edu/psu/cto5068/like_birbs/game/DisplayEventDialog.java
@@ -1,57 +1,0 @@
-package edu.psu.cto5068.like_birbs.game;
-
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.Context;
-import android.content.DialogInterface;
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.DialogFragment;
-
-public class DisplayEventDialog extends DialogFragment {
-
-    public interface EventDialogListener {
-        void onDialogPositiveClick(DialogFragment dialog);
-        void onDialogNegativeClick(DialogFragment dialog);
-    }
-
-    EventDialogListener listener;
-
-    @Override
-    public void onAttach(@NonNull Context context) {
-        super.onAttach(context);
-        try {
-            listener = (EventDialogListener) context;
-        } catch (ClassCastException e) {
-            throw new ClassCastException(this.toString() + " must implement EventDialogListener");
-        }
-    }
-
-    @NonNull
-    @Override
-    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        String title = getArguments().getString("title");
-        String message = getArguments().getString("message");
-
-        String posButton = getArguments().getString("posButton");
-        String negButton = getArguments().getString("negButton");
-
-        builder
-                .setTitle(title)
-                .setMessage(message)
-                .setPositiveButton(posButton, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        listener.onDialogPositiveClick(DisplayEventDialog.this);
-                    }
-                })
-                .setNegativeButton(negButton, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        listener.onDialogNegativeClick(DisplayEventDialog.this);
-                    }
-                });
-        return builder.create();
-    }
-}
