+++ app/src/main/java/ga/hnbenterprises/like_birbs/LogMsg.java
@@ -1,0 +1,17 @@
+package ga.hnbenterprises.like_birbs;
+
+public class LogMsg {
+    private String msg;
+
+    public LogMsg(String msg) {
+        this.msg = msg;
+    }
+
+    public String getMsg() {
+        return msg;
+    }
+
+    public void setMsg(String msg) {
+        this.msg = msg;
+    }
+}
