+++ app/src/main/java/ga/hnbenterprises/like_birbs/RandomEvent.java
@@ -1,0 +1,4 @@
+package ga.hnbenterprises.like_birbs;
+
+public class RandomEvent {
+}
