+++ app/src/main/java/ga/hnbenterprises/like_birbs/game/DisplayBirbs.java
@@ -1,0 +1,64 @@
+package ga.hnbenterprises.like_birbs.game;
+
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+
+import java.lang.reflect.Array;
+import java.util.ArrayList;
+import java.util.List;
+
+import ga.hnbenterprises.like_birbs.LogListAdapter;
+import ga.hnbenterprises.like_birbs.LogMsg;
+import ga.hnbenterprises.like_birbs.R;
+
+public class DisplayBirbs extends DialogFragment {
+
+    public interface LogDialogListener {
+        void onLogDialogPositiveClick(DialogFragment dialog);
+        void onLogDialogNegativeClick(DialogFragment dialog);
+    }
+
+    LogDialogListener listener;
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        try {
+            listener = (LogDialogListener) context;
+        } catch (ClassCastException e) {
+            throw new ClassCastException(this.toString() + " must implement LogDialogListener");
+        }
+    }
+
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+
+        ArrayList logs = getArguments().getStringArrayList("logs");
+
+        List<LogMsg> logMsgs = new ArrayList<>();
+        for (int i = 0; i < logs.size(); i++) {
+            logMsgs.add(new LogMsg(logs.get(i).toString()));
+        }
+
+        LayoutInflater inflater = requireActivity().getLayoutInflater();
+
+        LogListAdapter adapter = new LogListAdapter(getActivity(), R.layout.logs_adapter_view, logMsgs);
+
+        builder
+                .setTitle("CURRENT BIRBS IN COLONY")
+                .setView(inflater.inflate(R.layout.logs_view, null))
+                .setAdapter(adapter, null);
+
+        return builder.create();
+    }
+}
